% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aq.R
\name{aq}
\alias{aq}
\title{combined index value of all pollutants}
\usage{
aq(a, b, c, d, e, f, g, h)
}
\arguments{
\item{a, b, c, d, e, f, g, h}{numeric value of the individual pollutants(pm2.5,pm10,no,no2,nh3,so2,co,o3)}
}
\value{
print the air quality index
}
\description{
take concentration value of all pollutants
}
