% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co.R
\name{co}
\alias{co}
\title{sub-index of carbon monoxide(CO)}
\usage{
co(x)
}
\arguments{
\item{x}{the numeric value of Carbon monoxide in concentration(mg/m^3) 8-hr}
}
\value{
print the index of carbon monoxide (CO)
}
\description{
take concentration value (mg/m^3)of co and give the index value
}
