% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh3.R
\name{nh3}
\alias{nh3}
\title{sub-index of Ammonia (NH3)}
\usage{
nh3(x)
}
\arguments{
\item{x}{the concentration value of ammonia in ug/m^3}
}
\value{
print the index value of ammonia
}
\description{
take the concentration value(ug/m^3) of ammonia and give the index value
}
