% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nox.R
\name{nox}
\alias{nox}
\title{sub-index of NOx(any nitric x oxide:- NO2,NO)}
\usage{
nox(x)
}
\arguments{
\item{x}{the numerical value in concentration of NOx (ug/m^3)}
}
\value{
print the index value
}
\description{
take the concentration value (ug/m^3) and give the index
}
