% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o3.R
\name{o3}
\alias{o3}
\title{sub-index of ozone (O3)}
\usage{
o3(x)
}
\arguments{
\item{x}{concentration value of ozone in ug/m^3]}
}
\value{
the index value of ozone
}
\description{
take the concentration value of ozone in ug/m^3 8-hr
}
