% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm10.R
\name{pm10}
\alias{pm10}
\title{Sub-index of a PM10}
\usage{
pm10(x)
}
\arguments{
\item{x}{The numeric value in concentration(ug/m^3)}
}
\value{
The index value of parameter PM10
}
\description{
take the concentration value(ug/m^3) and give the index value
}
