% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm25.R
\name{pm2.5}
\alias{pm2.5}
\title{sub-index of PM2.5}
\usage{
pm2.5(x)
}
\arguments{
\item{x}{The numeric value in concentration (ug/m^3)}
}
\value{
the index value of the PM2.5
}
\description{
take the concentration value in (ug/m^3) and give the index value
}
