% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so2.R
\name{so2}
\alias{so2}
\title{sub-index of SO2}
\usage{
so2(x)
}
\arguments{
\item{x}{the numerical value of concentration(ug/m^3) in 24-hr}
}
\value{
the index value
}
\description{
take the concentration value (ug/m^3) and give the index
}
